--- a/.changeset/neat-pandas-explode.md
+++ b/.changeset/neat-pandas-explode.md
@@ -1,0 +1,5 @@
+---
+'@shopify/polaris': patch
+---
+
+Upgrade to Storybook v8