--- a/polaris-react/.storybook/test-runner.ts
+++ b/polaris-react/.storybook/test-runner.ts
@@ -1,0 +1,35 @@
+/* eslint-env node */
+import {injectAxe, checkA11y, configureAxe} from 'axe-playwright';
+import {getStoryContext} from '@storybook/test-runner';
+import type {TestRunnerConfig} from '@storybook/test-runner';
+
+/*
+ * See https://storybook.js.org/docs/writing-tests/test-runner#test-hook-api
+ * to learn more about the test-runner hooks API.
+ */
+const config: TestRunnerConfig = {
+  tags: {
+    skip: ['skip-tests'],
+  },
+  async preVisit(page) {
+    await injectAxe(page);
+  },
+  async postVisit(page, context) {
+    // Get the entire context of a story, including parameters, args, argTypes, etc.
+    const storyContext = await getStoryContext(page, context);
+
+    // Do not run a11y tests on disabled stories.
+    if (storyContext.parameters?.a11y?.disable) {
+      return;
+    }
+
+    // Apply story-level a11y rules
+    await configureAxe(page, {
+      rules: storyContext.parameters?.a11y?.config?.rules,
+    });
+
+    await checkA11y(page, '#storybook-root', {}, false, 'v2');
+  },
+};
+
+module.exports = config;