--- a/polaris-react/playground/Playground.tsx
+++ b/polaris-react/playground/Playground.tsx
@@ -3,6 +3,7 @@ import React from 'react';
 import {Page} from '../src';
 
 export const Playground = {
+  tags: ['skip-tests'],
   render() {
     return (
       <Page title="Playground">