--- a/polaris-react/src/components/ResourceList/ResourceList.stories.tsx
+++ b/polaris-react/src/components/ResourceList/ResourceList.stories.tsx
@@ -1520,6 +1520,7 @@ export const WithinAModal = {
       <Frame>
         <div style={{height: '500px'}}>
           <Modal
+            instant
             noScroll
             activator={activator}
             open={active}