… pnpm-lock.yaml
… polaris-react/package.json
… polaris-react/scripts/accessibility-check.js