About 90% of the upgrade was handled by `npx storybook@latest upgrade` 🎉  _(Feels like just [yesterday we did the upgrade from v6 -> v7](https://github.com/Shopify/polaris/pull/9019) 😅)_

## Node v18

Storybook v8 requires node v18. But we're still stuck on node v16.

🚨 This PR is blocked by;
- ~https://github.com/Shopify/polaris/pull/11844~
- ~https://github.com/Shopify/polaris/pull/11924~
- ~https://github.com/Shopify/polaris/pull/11930~

## Performance

Calculated using [hyperfine](https://github.com/sharkdp/hyperfine):

```sh
hyperfine --parameter-list branch main,storybook-8 --setup='git checkout {branch} && yarn && yarn build --filter=@shopify/polaris' --prepare='rm -rf polaris-react/build-internal/storybook/' --warmup 1 'yarn workspace @shopify/polaris run storybook:build --quiet'
```

**`main`**

```
Benchmark 1: yarn workspace @shopify/polaris run storybook:build --quiet
  Time (mean ± σ):     24.362 s ±  0.387 s    [User: 39.938 s, System: 4.076 s]
  Range (min … max):   23.896 s … 25.034 s    10 runs

```

**`storybook-8`**

```
Benchmark 2: yarn workspace @shopify/polaris run storybook:build --quiet
  Time (mean ± σ):     29.486 s ±  1.321 s    [User: 40.765 s, System: 3.868 s]
  Range (min … max):   28.419 s … 32.763 s    10 runs
```

**Results**

```
Summary
  yarn workspace @shopify/polaris run storybook:build --quiet (branch = main) ran
    1.21 ± 0.06 times faster than yarn workspace @shopify/polaris run storybook:build --quiet (branch = storybook-8)
```

Combined with the slow-down seen in [the upgrade to v7](https://github.com/Shopify/polaris/pull/9019), I'm starting to think Chromatic's marketing about better performance isn't true for us 😅